requirements.txt
azure-functions == 1.6.0
azure-storage-blob == 12.7.1
azure-storage-queue == 12.1.5